+++ macos/build-menubar.sh
@@ -52,7 +52,7 @@ cat > "$APP/Contents/Info.plist" <<'PLIST'
     <key>CFBundlePackageType</key>
     <string>APPL</string>
     <key>CFBundleShortVersionString</key>
-    <string>3.1.1</string>
+    <string>3.1.2</string>
     <key>CFBundleVersion</key>
     <string>311</string>
     <key>LSMinimumSystemVersion</key>